Updates launcher scripts for bundle paths
Updates the launcher scripts to correctly reference the bundled `pskWebServer.js` file. This ensures that the scripts can locate and execute the web server from the bundle directory.

Additionally, updates the `PSK_ROOT_INSTALATION_FOLDER` paths to point to the current directory, ensuring correct resolution of paths relative to the launcher scripts.